Check data quality module ongoing
